cambio del tipo de variable de size y index

+++ ft_bzero.c
@@ -6,16 +6,16 @@
 /*   By: alejhern <alejhern@student.42barcelon      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/12 22:23:23 by alejhern          #+#    #+#             */
-/*   Updated: 2024/06/12 22:28:26 by alejhern         ###   ########.fr       */
+/*   Updated: 2024/06/17 14:58:25 by alejhern         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_bzero(void *s, int n)
+void	ft_bzero(void *s, size_t n)
 {
 	unsigned char	*util;
-	int				index;
+	size_t			index;
 
 	util = (unsigned char *) s;
 	index = 0;
